Create arenalib with a stub executable
