+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "cSpell.words": [
+        "arenalib",
         "Ashlands",
         "ggez",
         "Hamons",
