+++ Cargo.toml
@@ -3,6 +3,10 @@ name = "arena-gs"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+crate-type = ["rlib", "dylib"]
+name = "arenalib"
+
 [dependencies]
 anyhow = "1.0.64"
 ggez = { path = "../ggez", default-features = false }
